+++ src/rfsuite/i18n/de.lua
@@ -1826,6 +1826,7 @@ return {
       waiting_for_receiver_telemetry = "Warte auf Empfänger-Telemetrie (1RSS/2RSS)",
       waiting_for_battery_telemetry = "Warte auf Batterie-Telemetrie",
       connected_starting = "Verbunden, Dashboard wird gestartet...",
+      connected_partial_telemetry = "Verbunden mit teilweiser Telemetrie",
       loading_tasks = "Lade Daten...",
       select_battery = "AKKU AUSWÄHLEN",
       switch_to_fullscreen = "Bitte in den Vollbildmodus wechseln um den Akku auszuwählen",